Moved handlers/ssl to plugin/connectors/tcp/connectors/ssl
Since this package is only used by pg/mysql TCP connectors, this handler
(which really wasn't a real handler) got moved to the same location where
they are.
